core: enable dumpable flag after setuid() on Linux
- init rlimit struct for proper dbg message when core limits are not
  changed
(cherry picked from commit f4136d7742b0b0bfc09a9157d573c8a7f6118b76)

+++ daemonize.h
@@ -30,7 +30,7 @@
 int daemonize(char* name, int daemon_status_fd_input);
 int do_suid();
 int increase_open_fds(int target);
-int set_core_dump(int enable, int size);
+int set_core_dump(int enable, long unsigned int size);
 int mem_lock_pages();
 int set_rt_prio(int prio, int policy);
